+++ src/jinja2/utils.py
@@ -3,7 +3,6 @@ import json
 import os
 import re
 import typing as t
-import warnings
 from collections import abc
 from collections import deque
 from random import choice
@@ -84,72 +83,7 @@ class _PassArg(enum.Enum):
         if hasattr(obj, "jinja_pass_arg"):
             return obj.jinja_pass_arg  # type: ignore
 
-        for prefix in "context", "eval_context", "environment":
-            squashed = prefix.replace("_", "")
-
-            for name in f"{squashed}function", f"{squashed}filter":
-                if getattr(obj, name, False) is True:
-                    warnings.warn(
-                        f"{name!r} is deprecated and will stop working"
-                        f" in Jinja 3.1. Use 'pass_{prefix}' instead.",
-                        DeprecationWarning,
-                        stacklevel=2,
-                    )
-                    return cls[prefix]
-
         return None
-
-
-def contextfunction(f: F) -> F:
-    """Pass the context as the first argument to the decorated function.
-
-    .. deprecated:: 3.0
-        Will be removed in Jinja 3.1. Use :func:`~jinja2.pass_context`
-        instead.
-    """
-    warnings.warn(
-        "'contextfunction' is renamed to 'pass_context', the old name"
-        " will be removed in Jinja 3.1.",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    return pass_context(f)
-
-
-def evalcontextfunction(f: F) -> F:
-    """Pass the eval context as the first argument to the decorated
-    function.
-
-    .. deprecated:: 3.0
-        Will be removed in Jinja 3.1. Use
-        :func:`~jinja2.pass_eval_context` instead.
-
-    .. versionadded:: 2.4
-    """
-    warnings.warn(
-        "'evalcontextfunction' is renamed to 'pass_eval_context', the"
-        " old name will be removed in Jinja 3.1.",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    return pass_eval_context(f)
-
-
-def environmentfunction(f: F) -> F:
-    """Pass the environment as the first argument to the decorated
-    function.
-
-    .. deprecated:: 3.0
-        Will be removed in Jinja 3.1. Use
-        :func:`~jinja2.pass_environment` instead.
-    """
-    warnings.warn(
-        "'environmentfunction' is renamed to 'pass_environment', the"
-        " old name will be removed in Jinja 3.1.",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    return pass_environment(f)
 
 
 def internalcode(f: F) -> F:
@@ -481,18 +415,6 @@ def url_quote(obj: t.Any, charset: str = "utf-8", for_qs: bool = False) -> str:
         rv = rv.replace("%20", "+")
 
     return rv
-
-
-def unicode_urlencode(obj: t.Any, charset: str = "utf-8", for_qs: bool = False) -> str:
-    import warnings
-
-    warnings.warn(
-        "'unicode_urlencode' has been renamed to 'url_quote'. The old"
-        " name will be removed in Jinja 3.1.",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    return url_quote(obj, charset=charset, for_qs=for_qs)
 
 
 @abc.MutableMapping.register
@@ -831,24 +753,3 @@ class Namespace:
 
     def __repr__(self) -> str:
         return f"<Namespace {self.__attrs!r}>"
-
-
-class Markup(markupsafe.Markup):
-    def __new__(cls, base="", encoding=None, errors="strict"):  # type: ignore
-        warnings.warn(
-            "'jinja2.Markup' is deprecated and will be removed in Jinja"
-            " 3.1. Import 'markupsafe.Markup' instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return super().__new__(cls, base, encoding, errors)
-
-
-def escape(s: t.Any) -> str:
-    warnings.warn(
-        "'jinja2.escape' is deprecated and will be removed in Jinja"
-        " 3.1. Import 'markupsafe.escape' instead.",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-    return markupsafe.escape(s)